+++ server/lessons/4l7NcEIYDgqzK3yR/evts.js
@@ -1,0 +1,6 @@
+import * as a from "../../../src/lesson";
+return [
+    new LE_AddGBubble([
+        "Hello!"
+    ])
+]