+++ server/lessons/PpcVdfjdzo7TfWBL/evts.js
@@ -1,0 +1,6 @@
+import * as a from "../../../src/lesson";
+return [
+    new LE_AddGBubble([
+        "I've brought over both our sets of files for this tutorial because we're going to add on to it."
+    ])
+]