+++ server/lessons/yX0zoNI7fjG0MgtQ/css_board.js
@@ -33,15 +33,15 @@ return [
     new BE_GlobalBubble(`This code changes the background color of the body of our page to darkcyan.\nLet's break down it's structure.`),
     new BE_SetOpacity([".","style1","val1"],0),
     new BE_Wait(300),
-    new BE_CircleObj(["selector"]).applyToObjs(o=>{
+    new BE_UnderlineObj(["selector"]).applyToObjs(o=>{
         o.addTag("_circle");
     }),
-    new BE_GlobalBubble(`The first part is called the h[selector.]\nIt specifies the type of elements we want to change here, and in this example we want to select our body element.`),
+    new BE_GlobalBubble(`The first part is called the h[selector.]\nIt specifies the type of elements we want to change here, and in this example we want to select our b[body] element.`),
     new BE_RemoveObj(["_circle"]),
     new BE_SetOpacity(["selector",0]),
     new BE_GlobalBubble(`Everything within these curly braces will be the i["styles"] that we add to those elements.`),
     new BE_SetOpacity([".","style1","val1"],1),
-    new BE_GlobalBubble(`All styles follow this same pattern:\ni[style-name: value;]`),
+    new BE_GlobalBubble(`All styles follow this same pattern:\ni[style-name: value;]\ni[And we instead of using spaces CSS uses dashes (-).]`),
     new BE_GlobalBubble(`This whole structure is called a h[Style Set], because it contains a set of properties or styles.`),
     // new BE_GlobalBubble(`This is the basic structure of `),